migration: add a basic migration script
As future versions are likely to introduce breaking changes in the
json files (new fields, ...). We need a way to update existing patchr
configurations to make them work with newer binaries.

This commit adds a simple Python script that converts json files
to make them compatible with the newest version of patchr.

migrations.py will be extended after every breaking change to make
the migration as smooth as possible.